services/extraction-agent/pyproject.toml
[project]
name = "extraction-agent"
version = "0.1.0"
description = "Extraction Agent (Agent 1) for Ancol MoM Compliance System"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "google-genai>=1.0",
    "google-cloud-storage>=2.0",
    "ancol-common",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/extraction_agent"]
